pyproject.toml
[project]
name = "claude-telegram-bot"
version = "0.1.0"
description = "Telegram bot that interfaces with Claude API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "telethon>=1.37.0",
    "anthropic>=0.39.0",
    "python-dotenv>=1.0.0",
    "cryptography>=43.0.0",
    "aiohttp>=3.9.0",
]